requirements.txt
asgiref==3.7.2
Django==4.2.4
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
packaging==23.1
phonenumberslite==8.13.19
Pillow==10.0.1
psycopg2==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.5.0
